Add support for poetry

+++ guix_env/cli.py
@@ -6,6 +6,7 @@ import tempfile
 import shutil
 from jinja2 import Environment, FileSystemLoader
 
+# TODO: add test that the environment exists before doing anything.
 
 main_dir=os.path.join(os.getenv("HOME"), ".guix_env")
 file_path = os.path.realpath(__file__)
@@ -13,14 +14,26 @@ file_path = os.path.realpath(__file__)
 environment = Environment(loader=FileSystemLoader(
     os.path.join(os.path.dirname(file_path),"template_scripts/")))
 
-default_guix_packages = [
+guix_python_packages = [
     "python",
     "python-toolchain",
+    "poetry-next", # this comes from perso channel while waiting for guix to have a newer version of poetry
+    "xcb-util", # xcb is for matplotlib to be able to plt.show
+    "xcb-util-wm",
+    "xcb-util-image",
+    "xcb-util-keysyms",
+    "xcb-util-renderutil",
+    "xcb-util-cursor",
+    ]
+
+default_guix_packages = [
     "bash",
     "glibc-locales",
     "nss-certs",
     "coreutils",
     "diffutils",
+    "findutils",
+    "curl",
     "git",
     "make",
     "zlib",
@@ -28,18 +41,12 @@ default_guix_packages = [
     "tcl",
     "gtk",
     "grep",
-    "xcb-util", # xcb/dbus is for matplotlib to be able to plt.show
-    "xcb-util-wm",
-    "xcb-util-image",
-    "xcb-util-keysyms",
-    "xcb-util-renderutil",
     "dbus",
-    "xcb-util-cursor",
     "ncurses",
+    "nano",
+    "tmux",
     "zsh"
 ]
-
-default_python = ["numpy", "matplotlib", "PyQt5", "scipy"]
 
 @click.group()
 @click.pass_context
@@ -53,83 +60,87 @@ def guix_env(ctx):
 @click.argument('name',required = True, type=str)
 @click.option('--channel-file',required = False, type=str, help="Path to a channel file to be used in the guix install")
 @click.option('--requirements-file',required = False, type=str, help="Path to a requirements.txt file to be used in the python install")
-@click.option('--manifest-file',required = False, type=str, help="Path to a manifest file to be used in the guix install")
+@click.option('--without-python', is_flag=True, help="Do an environment without python")
+@click.option('--pyproject-file',required = False, type=str, help="Path to a pyproject.toml file to be used in the python install (override requirement file if both are given).")
+@click.option('--poetry-lock-file',required = False, type=str, help="Path to a poetry.lock file to be used in the python install")
+@click.option('--manifest-file',required = False, type=str, help="Path to a manifest file to be used in the guix install. Will replace the default manifest.")
 @click.option('--guix-args',required = False, type=str, default="-CFNW", help="arguments to be passed to guix")
 @click.pass_context
-def create(ctx, name, channel_file, requirements_file, manifest_file, guix_args):
+def create(ctx, name, channel_file, without_python, requirements_file, pyproject_file, poetry_lock_file, manifest_file, guix_args):
     """
     Create an environment with name `name`. A channel file can be specified, otherwise a channel file will be
     automatically created.
     """
+    with_python = not without_python
     assert not os.path.isdir(os.path.join(main_dir, name)), "Environment already exist"
     os.system('mkdir -p '+os.path.join(main_dir, name, "bin"))
+    os.system('mkdir -p '+os.path.join(main_dir, name, ".local"))
+    home = os.getenv("HOME")
 
-    zshrc = environment.get_template("zshrc").render(name = name, reqfile = os.path.join(main_dir, name, "requirements.txt"))
+    zshrc = environment.get_template("zshrc").render(name = name, reqfile = os.path.join(main_dir, name, "requirements.txt"), with_python=with_python)
+    run_script = environment.get_template("run_script.sh").render(name=name, guix_args = guix_args, HOME=home, with_python=with_python)
 
-    # TBC
-    
-    if channel_file is None:
-        channels = subprocess.run(["guix", "describe", "-f", "channels"], capture_output=True).stdout.decode()
-    else:
-        channels = subprocess.run(["cat", channel_file], capture_output=True).stdout.decode()
-
-    template = environment.get_template("activate.sh")
-    script = template.render(name = name, guix_args = guix_args)
-
-    with open(os.path.join(main_dir, name, "bin", ".zshrc"), "w") as myfile:
-        myfile.write(zshrc)
-    
-    with open(os.path.join(main_dir, name, "bin", "activate.sh"), "w") as myfile:
-        myfile.write(script)
-        os.system('chmod +x '+os.path.join(main_dir, name, "bin", "activate.sh"))
+    channels = _make_channel_file(channel_file)
         
-    with open(os.path.join(main_dir, name, "bin", "run.sh"), "w") as myfile:
-
-        template_run = environment.get_template("run_script.sh")
-        run_script = template_run.render(name = name, guix_args = guix_args)
+    with open(os.path.join(main_dir, name, ".zshrc"), "w") as myfile:
+        myfile.write(zshrc)
+    with open(os.path.join(main_dir, name, "bin", "run_script.sh"), "w") as myfile:
         myfile.write(run_script)
-        os.system('chmod +x '+os.path.join(main_dir, name, "bin", "run.sh"))
-
+        os.system('chmod +x '+os.path.join(main_dir, name, "bin", "run_script.sh"))
+        
+    # Guix manifest and channel files
     if channel_file is None:
         with open(os.path.join(main_dir, name, "channels.scm"), "w") as myfile:
             myfile.write(channels)
     else:
         os.system("cp "+channel_file+" "+os.path.join(main_dir, name, "channels.scm"))
-
+        
     if manifest_file is None:
         with open(os.path.join(main_dir, name, "manifest.scm"), "w") as myfile:
             packages = default_guix_packages
+            if with_python:
+                packages = packages + guix_python_packages
             myfile.write(
                     "(specifications->manifest '(\n\"" + '"\n "'.join(packages) + '"\n))'
                 )
     else:
         os.system("cp "+manifest_file+" "+os.path.join(main_dir, name, "manifest.scm"))
-    
-    print("Creation of virtual environment")
-    os.system("guix time-machine --channels=${HOME}/.guix_env/"+name+"/channels.scm -- shell  python  -- python3 -m venv ~/.guix_env/guix_env_venv/"+name+"_venv")
-    run_file = os.path.join(main_dir, name, "bin", "run.sh")
-    if requirements_file is None:
-        print("Installing default python libs")
-        with open(os.path.join(main_dir, name, "requirements.txt"), "w") as myfile:
-            myfile.write("\n".join(default_python))
 
-        requirements_file = os.path.join(main_dir, name, "requirements.txt")
-    
-    os.system(run_file+" python -m pip install -r "+requirements_file)
-    os.system(run_file+" python -m pip freeze > "+os.path.join(main_dir, name, "requirements.txt"))
+    # 
+    with open(os.path.join(main_dir, name, "bin", "launch_in_guix.sh"), "w") as myfile:
+        launcher = environment.get_template("launch_in_guix.sh").render(name=name, guix_args = guix_args,)
+        myfile.write(launcher)
 
+    os.system("chmod +x "+os.path.join(main_dir, name, "bin", "launch_in_guix.sh"))
+
+    with open(os.path.join(main_dir, name, "bin",  "launch_shell.sh"), "w") as myfile:
+        launcher = environment.get_template("launch_shell.sh").render(name=name, with_python=with_python)
+        myfile.write(launcher)
+    os.system("chmod +x "+os.path.join(main_dir, name, "bin",  "launch_shell.sh"))
+
+    if with_python:
+        ### construct a poetry environment optionally with the specified requirements
+        _make_python_env(main_dir, name, pyproject_file, poetry_lock_file, requirements_file)
+        
+    print(f"Guix-env environment {name} has beenn created, its files can be found in {os.path.join(main_dir, name)}")
     
+
 @guix_env.command()
 @click.argument('name',required = True, type=str)
 @click.pass_context
 def update(ctx, name):
     """
-    Update the channel file (and as a consequence, it will update the packages managed by guix at next shell/run).
+    Update the channel file to the current guix channel file (and as a consequence, it will update the packages managed by guix at next shell/run).
+    
     """
-    channels = subprocess.run(["guix", "describe", "-f", "channels"], capture_output=True).stdout.decode()
+    print("Updating channel file")
+    channels = _make_channel_file(os.path.join(main_dir, name, "channels.scm"))
     with open(os.path.join(main_dir, name, "channels.scm"), "w") as myfile:
         myfile.write(channels)
-
+    if os.path.isfile(os.path.join(main_dir, name, "pyproject.toml")):
+        print("Found python install, updating")
+        _launch_cmd(name, "gep update")
+        
 
 @guix_env.command()
 @click.argument('name',required = True, type=str)
@@ -138,17 +149,18 @@ def rm(ctx, name):
     """
     Remove a guix-env environment.
     """
-    shutil.rmtree(os.path.join(main_dir, name))
-    shutil.rmtree(os.path.join(main_dir, "guix_env_venv", name+"_venv"))
-
+    if os.path.isdir(os.path.join(main_dir, name)):
+        print("Removing ", os.path.join(main_dir, name))
+        shutil.rmtree(os.path.join(main_dir, name))
 
 @guix_env.command()
 @click.argument('name',required = True, type=str)
 @click.argument('pkg',required = True, type=str)
 @click.pass_context
-def add(ctx, name, pkg):
+def add_guix(ctx, name, pkg):
     """
-    Add the package `pkg` to the environment named `name`.
+    Add the guix package `pkg` to the environment named `name`.
+    Warning: if you add a package from inside an environment, the package will not be available until you reconstruct the environment.
     """
     with open(os.path.join(main_dir, name, "manifest.scm"), "r") as myfile:
         packages = myfile.read().split("(")[2].split(")")[0]
@@ -161,6 +173,17 @@ def add(ctx, name, pkg):
         myfile.write(
                 "(specifications->manifest '(\n\"" + '"\n "'.join(packages) + '"\n))'
             )
+    print(f"Package {name} added to the manifest") 
+
+@guix_env.command()
+@click.argument('name',required = True, type=str)
+@click.argument('pkg',required = True, type=str)
+@click.pass_context
+def add_python(ctx, name, pkg):
+    """
+    Add the python package `pkg` to the environment named `name`.
+    """
+    _launch_cmd(name, f"gep add  {pkg}")
 
 @guix_env.command()
 @click.pass_context
@@ -178,8 +201,7 @@ def info(ctx, name):
     Get informations on environment with name `name`.
     """
     click.echo("Environment located in "+os.path.join(main_dir, name))
-    run_file = os.path.join(main_dir, name, "bin", "run.sh")
-    os.system(run_file+" guix describe")
+    _launch_cmd(name," guix describe")
 
     with open(os.path.join(main_dir, name, "manifest.scm"), "r") as myfile:
         packages = myfile.read().split("(")[2].split(")")[0]
@@ -191,7 +213,7 @@ def info(ctx, name):
     click.echo("\n".join(packages))
     click.echo("-"*10)
     click.echo("Installed python packages")
-    os.system(run_file+" pip3 freeze")
+    _launch_cmd(name, "gep run pip3 freeze")
 
 @guix_env.command()
 @click.argument('name',required = True, type=str)
@@ -202,23 +224,12 @@ def shell(ctx, name, tmux, cwd):
     """
     Open a shell in the environment with name `name`.
     """
+    assert os.path.isdir(os.path.join(main_dir, name)), "Environment does not exist"
+
+    print(f"Welcome to your guix-env environment: {name}")
+    print("To install python package, use 'gep add package_name'. gep is ann alias for poetry that install things at the right place.")
     
-    activation_file = os.path.join(main_dir, name, "bin", "activate.sh")
-
-    if tmux:
-        child = subprocess.run(["tmux","has-session", "-t",  "guix_env_"+name],capture_output=True,text=True)
-        rc = child.returncode
-        if rc != 0:
-            print("env not launched yet, launching now")
-            os.system("tmux new-session -d -s guix_env_"+name+" "+activation_file)
-
-        if cwd:
-            wd = os.getcwd()
-            os.system("tmux send-keys -t guix_env_"+name+" \" cd "+wd+ " && clear\" ENTER")
-            print("done cwd")
-        os.system("tmux attach -t guix_env_"+name)
-    else:
-        os.system(activation_file)
+    os.system(os.path.join(main_dir, name, "bin", "launch_in_guix.sh") + " " + os.path.join(main_dir, name, "bin", "launch_shell.sh"))
 
 @guix_env.command()
 @click.argument('name',required = True, type=str)
@@ -231,11 +242,16 @@ def run(ctx, name, cmd):
     Example of usage is
     guix-env run my_env "ls $HOME/"
     """
-    run_file = os.path.join(main_dir, name, "bin", "run.sh")
-    os.system(run_file+" "+cmd)
-        
+
+    _launch_cmd(name, cmd)
+
+
+def _launch_cmd(name, cmd):
+    os.system(os.path.join(main_dir, name, "bin", "launch_in_guix.sh")+ " " + os.path.join(main_dir, name, "bin", "run_script.sh") + " "  + cmd)
+    
   
 def _is_in_guix(pkg):
+    print("Checking that the package is indeed a guix package")
     output = subprocess.run(["guix", "search", pkg], capture_output=True).stdout.decode()
     output = output.split("name: ")
     names = [o.split("\n")[0] for o in output]
@@ -244,3 +260,45 @@ def _is_in_guix(pkg):
         if pkg == name.strip():
             res = True
     return res
+
+def _make_channel_file(channel_file=None):
+    if channel_file is None:
+        system_channels = subprocess.run(["guix", "describe", "-f", "channels"], capture_output=True).stdout.decode()
+    else:
+        system_channels = subprocess.run(["cat", channel_file], capture_output=True).stdout.decode()
+    
+    channels = environment.get_template("channels.scm").render(system_channels = system_channels)
+    return channels
+
+
+def _make_python_env(main_dir, name, pyproject_file, poetry_lock_file, requirements_file):
+        guix_python_cmd = f"guix time-machine --channels=$HOME/.guix_env/{name}/channels.scm -- shell python -- python3 --version | cut -d ' ' -f 2"
+        python_version = subprocess.check_output(guix_python_cmd, shell=True).decode().strip()
+
+        if pyproject_file is None:
+            author = subprocess.run(["whoami"], capture_output=True).stdout.decode()
+            pyproject = environment.get_template("pyproject.toml").render(name = name, python_version = python_version)
+        else:
+            with open(pyproject_file, "r") as myfile:
+                pyproject = myfile.read()
+
+        with open(os.path.join(main_dir, name,  "pyproject.toml"), "w") as myfile:
+            myfile.write(pyproject)
+
+        if poetry_lock_file is not None:
+            os.system(f"cp {poetry_lock_file} {os.path.join(main_dir, name)}")
+
+        if requirements_file is None:
+            reqfile = ""
+        else:
+            os.system("cp "+os.path.realpath(requirements_file)+ " /tmp/requirements_for_guix_env.txt")
+            reqfile = "/tmp/requirements_for_guix_env.txt"
+
+
+        create_env_file = environment.get_template("create_env.sh").render(name=name,
+                                                                           directory = os.path.join(main_dir, name),
+                                                                           requirements = reqfile)
+        with open(os.path.join("/tmp",  "create_guix_env.sh"), "w") as myfile:
+            myfile.write(create_env_file)
+        os.system("chmod +x "+os.path.join("/tmp",  "create_guix_env.sh"))
+        os.system(os.path.join(main_dir, name, "bin", "launch_in_guix.sh")+ " " + os.path.join("/tmp",  "create_guix_env.sh"))
